Move feature cohorts to `h.admin.features`
