+++ h/admin/views/__init__.py
@@ -6,7 +6,6 @@ def includeme(config):
     config.include('.admins')
     config.include('.badge')
     config.include('.features')
-    config.include('.cohorts')
     config.include('.groups')
     config.include('.nipsa')
     config.include('.staff')
