+++ h/admin/views/features.py
@@ -5,6 +5,7 @@ from pyramid import session
 from pyramid.view import view_config
 
 from h import models
+from h import paginator
 from h.i18n import TranslationString as _
 
 
@@ -31,6 +32,31 @@ def features_save(request):
     request.session.flash(_("Changes saved."), "success")
     return httpexceptions.HTTPSeeOther(
         location=request.route_url('admin_features'))
+
+
+@view_config(route_name='admin_cohorts',
+             request_method='GET',
+             renderer='h:templates/admin/cohorts.html.jinja2',
+             permission='admin_features')
+@paginator.paginate
+def cohorts_index(context, request):
+    query = request.db.query(models.FeatureCohort)
+    return query.order_by(models.FeatureCohort.name)
+
+
+@view_config(route_name='admin_cohorts',
+             request_method='POST',
+             request_param='add',
+             renderer='h:templates/admin/cohorts.html.jinja2',
+             permission='admin_features')
+def cohorts_add(request):
+    """Create a new feature cohort."""
+    cohort_name = request.params['add']
+    cohort = models.FeatureCohort(name=cohort_name)
+    request.db.add(cohort)
+
+    url = request.route_url('admin_cohorts')
+    return httpexceptions.HTTPSeeOther(url)
 
 
 def includeme(config):
