+++ tests/h/admin/views/features_test.py
@@ -1,8 +1,10 @@
 # -*- coding: utf-8 -*-
 
+from pyramid import httpexceptions
 from pyramid.testing import DummyRequest
 import pytest
 
+from h import db
 from h.admin.views import features as views
 
 
@@ -67,9 +69,27 @@ def test_features_save_checks_csrf_token(Feature, check_csrf_token):
     check_csrf_token.assert_called_with(request)
 
 
+def test_cohorts_index_when_no_cohorts():
+    req = DummyRequest(db=db.Session)
+    result = views.cohorts_index({}, req)
+    assert result["results"] == []
+
+
+def test_new_cohort_creates_cohort():
+    req = DummyRequest(db=db.Session)
+    req.params['add'] = 'cohort'
+    result = views.cohorts_add(req)
+    assert isinstance(result, httpexceptions.HTTPSeeOther)
+
+    result = views.cohorts_index({}, req)
+    assert len(result["results"]) == 1
+    assert result["results"][0].name == "cohort"
+
+
 @pytest.fixture(autouse=True)
 def routes(config):
     config.add_route('admin_features', '/adm/features')
+    config.add_route('admin_cohorts', '/adm/cohorts')
 
 
 @pytest.fixture
